delete bad pagination, set up new, change some styles
