--- src/components/CharacterList/CharacterList.styled.tsx
@@ -1,1 +1,0 @@
-import styled from 'styled-components/native';
