+++ src/components/CharacterList/CharacterList.styles.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components/native';
+
+export const CharasterListContainer = styled.View`
+  height: 100%;
+`;
