+++ src/screens/ScreeStarShips/ScreenStarShips.styles.tsx
@@ -3,6 +3,6 @@ import {Dimensions} from 'react-native';
 
 export const ScreenContainer = styled.View`
   width: ${Dimensions.get('window').width}px;
-  height: ${Dimensions.get('window').height}px;
+  height: 100%;
 `;
 //
