+++ src/screens/ScreenVehicles/ScreenVehicles.styles.ts
@@ -3,5 +3,5 @@ import {Dimensions} from 'react-native';
 
 export const VehieclesContainer = styled.View`
   width: ${Dimensions.get('window').width}px;
-  height: ${Dimensions.get('window').height}px;
+  height: 100%;
 `;
